Remove Auth0.js
Remove Auth0.js and redirect to /authorize from the server instead.
