+++ examples/auth0-aspnet-mvc4-sample/aspnet4-sample1/Views/Home/Index.cshtml
@@ -7,13 +7,13 @@
     <h1>ASP.NET with Auth0 Lock</h1>
     <p class="lead">ASP.NET is a free web framework for building great Web sites and Web applications using HTML, CSS and JavaScript.</p>
     
-    <p class="lead">This sample app includes <a href="https://auth0.com/docs/libraries/auth0js/v8" target="_blank" class="alert-link">Auth0.js</a>. Click the button below to try it out and watch for the protected pages you can access through the navbar at the top once you authenticate!</p>
     @if (User.Identity.IsAuthenticated)
     {
         <p>Welcome, @ViewBag.Name!</p>
-        <p><a href="/Home/Logout" class="btn btn-primary btn-lg">Logout</a></p>
+        <p><a href="/Account/Logout" class="btn btn-primary btn-lg">Logout</a></p>
     }
-    else {
-        <p><a href="#" class="btn btn-primary btn-lg" onclick="window.signin();">Login with Auth0</a></p>
+    else
+    {
+        <p>@Html.ActionLink("Login with Auth0", "Login", "Account", null, new { @class = "btn btn-primary btn-lg" })</p>
     }
 </div>
